+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_BASE_ENDPOINT = https://minoffinance-pro.onrender.com
+REACT_APP_BASE_ENDPOINT = https://minoffinance-hck7.onrender.com
