+++ src/containers/organizationonboardingarea/index.js
@@ -1,0 +1,12 @@
+import { Dashboard } from "../dashboard"
+import { OrganizationOnboardingAreaWrapper } from "./styled"
+
+export const OrganizationOnboardingArea = () => {
+    return (
+        <Dashboard>
+            <OrganizationOnboardingAreaWrapper>
+                <h1>Organization onboarding area</h1>
+            </OrganizationOnboardingAreaWrapper>
+        </Dashboard>
+    )
+}
