+++ src/containers/useronboardingarea/index.js
@@ -1,0 +1,12 @@
+import { Dashboard } from "../dashboard"
+import { UserOnboardingAreaWrapper } from "./styled"
+
+export const UserOnboardingArea = () => {
+    return (
+        <Dashboard>
+            <UserOnboardingAreaWrapper>
+                User onboarding to application area
+            </UserOnboardingAreaWrapper>
+        </Dashboard>
+    )
+}
